Make many ClusterController fields private
